refactor: merge createCountProxy into createDispatchProxy
