+++ src/api/rest/proxies/index.js
@@ -1,2 +1,1 @@
 export * from './create-dispatch-proxy';
-export * from './create-count-proxy';
